Added working mergeChangesIntoBranch method in fancier implementation. Updated fuzzer to use it. Tests pass!

+++ src/merge.ts
@@ -2,9 +2,9 @@
 import * as causalGraph from "./causal-graph.js"
 import { ID, ListFugueSimple } from "./list-fugue-simple.js"
 import { ListOp, ListOpLog, ListOpType } from "./oplog.js"
-import { Branch, LVRange } from "./types.js"
-import { assertEq } from './utils.js'
-import assert from 'node:assert/strict'
+import { Branch, LV, LVRange } from "./types.js"
+import { assert, assertEq } from './utils.js'
+import {deepEqual} from 'node:assert/strict'
 
 enum ItemState {
   NotYetInserted = -1,
@@ -53,7 +53,8 @@ interface EditContext {
 }
 
 interface WalkItem {
-  // These ranges are half-open ranges of opIds. So they include the start, but not the end.
+  // These ranges are ranges of opIds. They're half-open - so they include the start,
+  // but not the end.
   retreat: [number, number][],
   advance: [number, number][],
   consume: [number, number],
@@ -243,7 +244,7 @@ function integrate(ctx: EditContext, cg: causalGraph.CausalGraph, newItem: Item,
   // We've found the position. Insert where the cursor points.
 }
 
-function apply1<T>(ctx: EditContext, dest: T[], oplog: ListOpLog<T>, opId: number, fugue: ListFugueSimple<T>) {
+function apply1<T>(ctx: EditContext, dest: T[] | null, oplog: ListOpLog<T>, opId: number, fugue?: ListFugueSimple<T>) {
   const opIdToFugueId = (id: number): ID | null => {
     if (id === -1) return null
     const rawId = causalGraph.lvToRaw(oplog.cg, id)
@@ -274,7 +275,7 @@ function apply1<T>(ctx: EditContext, dest: T[], oplog: ListOpLog<T>, opId: numbe
 
     // Delete it in the output.
     if (item.endState === ItemState.Inserted) {
-      dest.splice(cursor.endPos, 1)
+      if (dest) dest.splice(cursor.endPos, 1)
     }
 
     // And mark the item as deleted. For the curState, we can't get into a "double deletes"
@@ -285,10 +286,12 @@ function apply1<T>(ctx: EditContext, dest: T[], oplog: ListOpLog<T>, opId: numbe
     // And mark that this delete corresponds to *that* item.
     ctx.delTargets[opId] = item.opId
 
-    fugue.receivePrimitive({
-      type: 'delete',
-      id: opIdToFugueId(item.opId)!,
-    })
+    if (fugue != null) {
+      fugue.receivePrimitive({
+        type: 'delete',
+        id: opIdToFugueId(item.opId)!,
+      })
+    }
   } else {
     // Insert! This is much more complicated as we need to do the Yjs integration.
     const cursor = findByCurPos(ctx, op.pos)
@@ -324,6 +327,8 @@ function apply1<T>(ctx: EditContext, dest: T[], oplog: ListOpLog<T>, opId: numbe
       originLeft,
       rightParent,
     }
+    if (ctx.itemsByLV[opId] != null) debugger
+    assert(ctx.itemsByLV[opId] == null, 'Inserted item already in list')
     ctx.itemsByLV[opId] = newItem
 
     // This will update the cursor to find the location we want to insert the item.
@@ -332,18 +337,20 @@ function apply1<T>(ctx: EditContext, dest: T[], oplog: ListOpLog<T>, opId: numbe
     ctx.items.splice(cursor.idx, 0, newItem)
 
     // And finally, actually insert it in the resulting document.
-    dest.splice(cursor.endPos, 0, op.content!)
+    if (dest) dest.splice(cursor.endPos, 0, op.content!)
 
-    let leftId = opIdToFugueId(originLeft) ?? fugue.start.id
-    let rightId = opIdToFugueId(tempOriginRight) ?? fugue.end.id
-    // console.log('inserting', opIdToFugueId(opId), 'left', leftId, 'right', rightId)
-    fugue.receivePrimitive({
-      type: 'insert',
-      id: opIdToFugueId(opId)!,
-      leftOrigin: leftId,
-      rightOrigin: rightId,
-      value: op.content!
-    })
+    if (fugue != null) {
+      let leftId = opIdToFugueId(originLeft) ?? fugue.start.id
+      let rightId = opIdToFugueId(tempOriginRight) ?? fugue.end.id
+      // console.log('inserting', opIdToFugueId(opId), 'left', leftId, 'right', rightId)
+      fugue.receivePrimitive({
+        type: 'insert',
+        id: opIdToFugueId(opId)!,
+        leftOrigin: leftId,
+        rightOrigin: rightId,
+        value: op.content!
+      })
+    }
   }
 }
 
@@ -377,17 +384,21 @@ function debugPrintCtx<T>(ctx: EditContext, oplog: ListOpLog<T>) {
   }
 }
 
-export function checkoutSimple<T>(oplog: ListOpLog<T>): Branch<T> {
-  const fugue = new ListFugueSimple<T>('_unused_')
+function walkBetween<T>(ctx: EditContext, oplog: ListOpLog<T>, data: T[] | null, fugue?: ListFugueSimple<T>) {
+  return walkBetween2(ctx, oplog, [], 0, causalGraph.nextLV(oplog.cg), data, fugue)
+}
 
-  const ctx: EditContext = {
-    items: [],
-    delTargets: new Array(oplog.ops.length).fill(-1),
-    itemsByLV: new Array(oplog.ops.length).fill(null),
-  }
 
-  const data: T[] = []
-  for (const {retreat, advance, consume} of walkCG(oplog.cg)) {
+function walkBetween2<T>(ctx: EditContext, oplog: ListOpLog<T>, curLV: number[], vStart: number, vEnd: number, data: T[] | null, fugue?: ListFugueSimple<T>): LV[] {
+  if (data != null) assert(data.length <= ctx.items.length)
+
+  // This function also needs to return the resulting version after doing this
+  // walk. (Which will be [vEnd - 1], but still).
+  const gen = walkCG(oplog.cg, curLV, vStart, vEnd)
+
+  let entry
+  while (!(entry = gen.next()).done) {
+    const {retreat, advance, consume} = entry.value
     // console.log('r', retreat, 'a', advance, 'c', consume)
 
     // Retreat.
@@ -415,6 +426,20 @@ export function checkoutSimple<T>(oplog: ListOpLog<T>): Branch<T> {
       apply1(ctx, data, oplog, lv, fugue)
     }
   }
+  return entry.value
+}
+
+export function checkoutSimple<T>(oplog: ListOpLog<T>): Branch<T> {
+  const fugue = new ListFugueSimple<T>('_unused_')
+
+  const ctx: EditContext = {
+    items: [],
+    delTargets: new Array(oplog.ops.length).fill(-1),
+    itemsByLV: new Array(oplog.ops.length).fill(null),
+  }
+
+  const data: T[] = []
+  walkBetween(ctx, oplog, data, fugue)
 
   // Deep weird check.
   // const expectedData = ctx.items
@@ -426,7 +451,7 @@ export function checkoutSimple<T>(oplog: ListOpLog<T>): Branch<T> {
 
   const fugueState = fugue.toArray()
   try {
-    assert.deepEqual(fugueState, data)
+    deepEqual(fugueState, data)
   } catch (e) {
     fugue.debugPrint()
     console.log()
@@ -435,101 +460,104 @@ export function checkoutSimple<T>(oplog: ListOpLog<T>): Branch<T> {
     throw e
   }
 
+  // console.log(ctx.items)
+  // fugue.debugPrint()
+  // console.log()
+  // debugPrintCtx(ctx, oplog)
+
   return { data, version: oplog.cg.heads }
 }
 
-// export function mergeLatest<T>(branch: Branch<T>, oplog: ListOpLog<T>) {
-//   // We have an existing checkout of a list document. We want to merge some new changes in the oplog into
-//   // our local branch.
-//   //
-//   // How do we do that?
-//   //
-//   // Obviously we could regenerate the branch from scratch - but:
-//   // - That would be very slow
-//   // - It would require reading all the old (existing) list operations - which we want to leave on disk
-//   // - That approach wouldn't allow us to get the difference between old and new state, which is
-//   //   important for updating cursor information and things like that.
-//   //
-//   // Ideally we only want to look at the new operations in the oplog and apply those. However,
-//   // those new operations may be concurrent with operations we've already processed. In that case,
-//   // we need to populate the items list with any potentially concurrent items.
-//   //
-//   // The strategy here looks like this:
-//   // 1. Find the most recent common ancestor of the existing branch & changes we're merging
-//   // 2. Re-iterate through the "conflicting set" of changes we've already merged to populate the items list
-//   // 3. Process the new operations as normal, starting with the crdt items list we've just generated.
+export function mergeChangesIntoBranch<T>(branch: Branch<T>, oplog: ListOpLog<T>, mergeVersion: number[] = oplog.cg.heads) {
+  // We have an existing checkout of a list document. We want to merge some new changes in the oplog into
+  // our local branch.
+  //
+  // How do we do that?
+  //
+  // Obviously we could regenerate the branch from scratch - but:
+  // - That would be very slow
+  // - It would require reading all the old (existing) list operations - which we want to leave on disk
+  // - That approach wouldn't allow us to get the difference between old and new state, which is
+  //   important for updating cursor information and things like that.
+  //
+  // Ideally we only want to look at the new operations in the oplog and apply those. However,
+  // those new operations may be concurrent with operations we've already processed. In that case,
+  // we need to populate the items list with any potentially concurrent items.
+  //
+  // The strategy here looks like this:
+  // 1. Find the most recent common ancestor of the existing branch & changes we're merging
+  // 2. Re-iterate through the "conflicting set" of changes we've already merged to populate the items list
+  // 3. Process the new operations as normal, starting with the crdt items list we've just generated.
 
-//   // First lets see what we've got. I'll divide the conflicting range into two groups:
-//   // - The conflict set. (Stuff we've already processed that we need to process again).
-//   // - The new operations we need to merge
-//   const newOps: LVRange[] = []
-//   const conflictOps: LVRange[] = []
+  // First lets see what we've got. I'll divide the conflicting range into two groups:
+  // - The conflict set. (Stuff we've already processed that we need to process again).
+  // - The new operations we need to merge
+  const newOps: LVRange[] = []
+  const conflictOps: LVRange[] = []
 
-//   let commonAncestor = causalGraph.findConflicting(oplog.cg, branch.version, oplog.cg.heads, (span, flag) => {
-//       // Note we'll be visiting these operations in reverse order.
-//       const target = flag === causalGraph.DiffFlag.B ? newOps : conflictOps
-//       target.push(span)
-//   })
-//   // newOps and conflictOps will be filled in in reverse order. Fix!
-//   newOps.reverse(); conflictOps.reverse()
+  let commonAncestor = causalGraph.findConflicting(oplog.cg, branch.version, mergeVersion, (span, flag) => {
+      // Note this visitor function visits these operations in reverse order.
+      const target = flag === causalGraph.DiffFlag.B ? newOps : conflictOps
+      // target.push(span)
 
-//   const ctx: EditContext = {
-//     items: [],
-//     delTargets: new Array(oplog.ops.length).fill(-1),
-//     itemsByLV: new Array(oplog.ops.length).fill(null),
-//   }
-//   // We need some placeholder items to correspond to the document as it looked at the commonAncestor state.
-//   // The placeholderLength just needs to be bigger than the document was at the time. This is inefficient but simple.
-//   let conflictOpLen = conflictOps.reduce((sum, [start, end]) => sum + end - start, 0)
-//   const placeholderLength = branch.data.length + conflictOpLen
-//   assert(placeholderLength <= Math.min(...commonAncestor))
+      let last
+      if (target.length > 0 && (last = target[target.length - 1])[0] === span[1]) {
+        last[0] = span[0]
+      } else {
+        target.push(span)
+      }
+  })
+  // newOps and conflictOps will be filled in in reverse order. Fix!
+  newOps.reverse(); conflictOps.reverse()
 
-//   for (let i = 0; i < placeholderLength; i++) {
-//     const item: Item = {
-//       opId: i, // TODO: Consider using some weird IDs here instead of normal numbers to make it clear.
-//       curState: ItemState.Inserted,
-//       endState: ItemState.Inserted,
-//       originLeft: -1,
-//       originRight: -1,
-//     }
-//     ctx.items.push(item)
-//     ctx.itemsByLV[item.opId] = item
-//   }
+  const ctx: EditContext = {
+    items: [],
+    delTargets: new Array(oplog.ops.length).fill(-1),
+    itemsByLV: new Array(oplog.ops.length).fill(null),
+  }
 
-//   for (const [start, end] of conflictOps) {
-//     // ... Walk.
-//   }
+  // We need some placeholder items to correspond to the document as it looked at the commonAncestor state.
+  // The placeholderLength needs to be at least the size that the document was at the time. This is inefficient but simple.
+  // let conflictOpLen = conflictOps.reduce((sum, [start, end]) => sum + end - start, 0)
+  // const placeholderLength = branch.data.length + conflictOpLen
+  const placeholderLength = Math.max(...branch.version) + 1
+  // const placeholderLength = Math.max(...commonAncestor)
+  // Also we must not use IDs that will show up in the actual document.
+  // assert(placeholderLength <= Math.min(...commonAncestor))
 
-//   for (const {retreat, advance, consume} of walkCG(oplog.cg, commonAncestor)) {
-//     // Retreat.
-//     // Note we're processing these in reverse order to make sure items
-//     // are undeleted before being un-inserted.
-//     for (let i = retreat.length - 1; i >= 0; i--) {
-//       const [start, end] = retreat[i]
-//       for (let lv = end - 1; lv >= start; lv--) {
-//         // console.log('retreat1', lv, oplog.ops[lv].type)
-//         retreat1(ctx, oplog, lv)
-//       }
-//     }
+  for (let i = 0; i < placeholderLength; i++) {
+    const opId = i + 1e12
+    const item: Item = {
+      // TODO: Consider using some weird IDs here instead of normal numbers to make it clear.
+      // Right now these IDs are also used in ctx.itemsByLV, but if that becomes a Map instead it would work better.
+      opId,
+      curState: ItemState.Inserted,
+      endState: ItemState.Inserted,
+      originLeft: -1,
+      rightParent: -1,
+    }
+    ctx.items.push(item)
+    ctx.itemsByLV[opId] = item
+  }
 
-//     // Advance.
-//     for (const [start, end] of advance) {
-//       for (let lv = start; lv < end; lv++) {
-//         advance1(ctx, oplog, lv)
-//       }
-//     }
+  let ctxVersion = commonAncestor
+  for (const [start, end] of conflictOps) {
+    // console.log('conflict', start, end)
+    // While processing the conflicting ops, we don't pass the document state because we don't
+    // want the document to be modified yet. We're just building up the items in ctx.
+    ctxVersion = walkBetween2(ctx, oplog, ctxVersion, start, end, null)
+  }
 
-//     // Then apply the operation.
-//     const [start, end] = consume
-//     for (let lv = start; lv < end; lv++) {
-//       // console.log('apply1', lv, oplog.ops[lv].type)
-//       apply1(ctx, branch.data, oplog, lv)
-//     }
-//   }
+  for (const [start, end] of newOps) {
+    // console.log('newOps', start, end)
+    // And now we update the branch.
+    ctxVersion = walkBetween2(ctx, oplog, ctxVersion, start, end, branch.data)
+  }
 
-//   // Set the branch version to the latest version.
-//   branch.version = oplog.cg.heads.slice()
-// }
+  // Set the branch version to the union of the versions.
+  // We can't use ctxVersion since it will probably just be the last version we visited.
+  branch.version = causalGraph.findDominators(oplog.cg, [...branch.version, ...mergeVersion])
+}
 
 export function mergeString(oplog: ListOpLog<string>): string {
   return checkoutSimple(oplog).data.join('')
